backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0

# Validation and serialization
pydantic[email]==2.5.0

# File handling and storage
aiofiles==23.2.0
boto3==1.34.0
python-magic==0.4.27

# Background tasks
celery[redis]==5.3.4

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# Monitoring and logging
structlog==23.2.0

# Environment management
python-dotenv==1.0.0

# CORS and security
# FastAPI has built-in CORS middleware

# Optional: for advanced features
redis==4.6.0
pillow==10.1.0